Upstreaming 1.5: Abandon failed sweep conditions in Lazarus experiment
Summary: For consistency of AE / Lazarus experiment with a sweep, we mark conditions that correspond to failed sweep runs as abandoned. This also helps in the following case of generation strategy: if a generation strategy uses gen. A for first 5 conditions and gen. B for second 5 conditions, but 3 out of 5 first conditions resulted in failed runs, gen. A should be used to propose some more conditions for the sweep. We achieve this functionality by adding an `exclude_abandoned` argument to generation strategy.

Differential Revision: D14024774

fbshipit-source-id: bf313032d3a859934abbe2713ca84a6e753e594a