1/n Creating FB-only SQA classes
Summary:
Splitting up D14083994 (storage for AnalysisSchedulers and AnalysisRuns) into 4 smaller diffs.

(This is all blocking for shadow reports, so would be great to get in soonish.)

This diff lays the groundwork by creating FB-specific versions of our SQL alchemy classes. This is necessary because we don't want to open source any of the analysis-related stuff. I originally thought we could just separate those two classes, but because of SQL relationships, it looks like we need to copy everything.

Reviewed By: amurthy1

Differential Revision: D14203973

fbshipit-source-id: 0a7710ec26f83b370a5f3925a777e0675311b425