4/n Saving/loading AnalysisRun and AnalysisScheduler
Summary:
Splitting up D14083994 (storage for AnalysisSchedulers and AnalysisRuns) into 4 smaller diffs.

This is the last one, and adds encoding/decoding and saving/loading.

Most of this is straightforward but tedious, i.e. encoding/decoding all the params, metrics, and constraints associated with the scheduler, just like we do for those that are attached to an experiment or trial.

Some of the save/load logic is a little different, since I'm creating new functions for saving/loading analysis schedulers/runs independent of saving/loading experiments (these are the only classes we have that separate functionality for).

Reviewed By: amurthy1

Differential Revision: D14204142

fbshipit-source-id: 54a064b129b0fe92407a009cd0da5b2173bd4ab7