Move testing utils to fb
Summary:
Move checks for Pyre strict, full annotations, and runner for testloop to lazarus/fb. This is currently designed to work with FB infra and the custom buck macros we have internally.

We'll want to add checks for full annotations in open source too, but need to figure out how still (maybe as a pytest plugin?).

Reviewed By: amurthy1

Differential Revision: D14295720

fbshipit-source-id: 53120a95535e7205eb0d19ba813c8bee085275f8