Use force_init=True for storage tests
Summary:
Our storage tests rely on the fact that the database is empty at the start of each one -- otherwise, all sorts of validation fail in our OSS tests (https://travis-ci.com/facebook/Adaptive-Experiment).

(Not sure why this doesn't happen in buck -- it might do some sort of teardown automatically?)

To fix this, we either need to:
1) Add a tearDown method that drops all tables, and add a line in setUp that recreates them
2) Set force_init=True, which disposes the current session and creates a new one (with new tables)

I went with #2 cause it was simpler, but if this isn't the right use of force_init, I can change.

Reviewed By: amurthy1

Differential Revision: D14303357

fbshipit-source-id: 82e4e6a337f60e74efeb93e4cd86a04785870355