Don't perform an update if field hasn't changed
Summary:
We store timestamps as integers in the database, so we don't store microsecond precision. However, datetime objects in Python do have microsecond precision.

Thus, we can get into an interesting scenario: we haven't changed a datetime attribute at all, but SQLAlchemy thinks it has changed because it sees the microseconds on the Python object, which don't match the object in the session.

So then SQLAlchemy goes to perform an update on the row in the database, but the update fails saying "expected 1 row to be updated but 0 were", because somewhere along the way the microseconds are dropped and so nothing has actually changed in the database.

To fix this, let's explicitly check for equality before performing an update. Usually SQLAlchemy is smart enough to do this itself, but seems like this is at least one exception.

Reviewed By: kkashin

Differential Revision: D14322830

fbshipit-source-id: a25b1b4800ef6391a3abdd3df531ce5892d2e066