Keep nulls from old AE parameters when converting
Summary:
Old AE experiments allow `null` values for the status quo, primarily for code defaults specified in QE.

Lazarus enforces that all arms have all parameters defined.

When we convert now, we make sure that each arm has all parameters defined (even if the value must be imputed as `None`).

When we fit models, we now drop all parameters with null values. This leads to the model treating those arms as out of design: diffusion/FBS/browse/master/fbcode/ae/lazarus/ae/generator/array.py$333-336

`LabelBinarizer` should, in theory, handle categorical float labels. However, sklearn types these as `continuous`, and the binarizer breaks in an opaque way. So, I've added the float->int `LabelEncoder` as part of our `OneHot` transform.

I plan to submit a pull request to SKLearn to fix this, since it seems easy to support, and the docs do not suggest that it is unsupported.

Reviewed By: amurthy1

Differential Revision: D14320413

fbshipit-source-id: fcb81a973480fed06cc0feacf4c23dfe51ab9da1