Proper type casting/checking when assigning conditions to a batch/experiment
Summary:
When adding conditions or status quo to batch or experiment, we should have a centralized way to validate the conditions. I made 3 validate/cast functions which live on the search space.

1. check_membership - check if the condition is in the search space domains and satisfies parameter constraints
2. check_types - check that the given condition's parameter set matches and the types are the same.
3. cast - Throw if there is a parameter set mismatch, and cast each value to the type given by the search space.

I think all 3 are useful in different places but unclear which should be used when. Left some inline comments would love to have discussion.

Reviewed By: ldworkin

Differential Revision: D13552057

fbshipit-source-id: dec74ec8d8f64ba3e18c4c45ee423295fe110d04