Fix to docstrings to make Sphinx happy
Summary:
Fixing some docstrings to make Sphinx / autodocs happy.

A few important things to point out here (that I will compile into a guide for our docstrings).

Wrapped new lines for args must have a tab indent. The wrapped line doesn't need to line up with the start

For example:
```
Args:
    x_int: this is an integer
        with a long description
        that takes multiple lines
    y_int: this is another integer
```

For manual lists (not for args), leave a blank newline before and after the list. Do not indent the list + make sure any wrapped lines start at the exact same place as previous text (do not tab indent).

```
This is some text.

- First bullet.
- Second bullet
  that takes multiple lines
- Third bullet

Normal text again.
```

This has to be on a newline and indented; otherwise it won't show up in a special section. Also, you cannot use `:` within the description; else the type will not populate.

Don't do this:
```
Returns: x: an integer
```
or this
```
Returns:
    x: an integer
```

Do this:
```
Returns:
    An integer
```

For tuple returns (multiple values), do this (not that cannot put `:` after "A tuple containing"):

```
    Returns:
        A tuple containing

        - Mean prediction.
        - Variance prediction.
        - Gradient of f.
        - Gradient of f_var.
```

Reviewed By: ldworkin

Differential Revision: D14400528

fbshipit-source-id: a16435077f37d28aa80cb5fdec97b0b8beda05ef