Remove ABC from SQA base
Summary:
Cannot properly mock SQA with an abstract base class in the `declarative_base` (see D14400531 for mocking; this is needed for autodocs to work on SQA).

We don't really need the abstract metaclass anyway. This was useful in AE v0 where user-facing classes we the same as SQA classes and we needed to add an abstract base class as a metaclass so that we could define abstract methods on the user-facing classes. We don't have this use anymore, since SQA classes aren't user-facing and don't define any abstract methods.

Reviewed By: ldworkin

Differential Revision: D14400674

fbshipit-source-id: de62ff0692e53371183ff0e63dd073f8ff3f6026