Add scripts for embedding Sphinx in Docusaurus
Summary:
Add script to parse html output of Sphinx (extract HTML body), wrap it in a div, and move it to Docusaurus.

From https://fb.workplace.com/groups/docusaurus/permalink/758687551170035/?comment_id=762953987410058&comment_tracking=%7B%22tn%22%3A%22R%22%7D, what I'm doing here is:

1) Run Sphinx to generate full HTML pages
2) Extract the body of HTML (BeautifulSoup or something like that)
3) Place the body into pages/api subdirectory of Docusarus + make sure that wrapPagesHTML flag is set to true in the config
4) Package the CSS / JS from Sphinx with Docusaurus to get the proper styling. I'll need to adjust this to make it fit within the theme of the rest of the site.

Of course, this is only a partial solution since this will basically be a standalone full API reference; if I reference a class or module in a tutorial I'm writing within Docusaurus, it would be great to have it link to the proper documentation in Sphinx, but that's not a v0.

Reviewed By: Balandat

Differential Revision: D14402186

fbshipit-source-id: bd48a4af41081ba1932e354aa106838814285000