Move `GenerationStrategy` out of `benchmark` and into `generator` + make `filter_kwargs` private and move into `GenerationStrategy`
Summary: `GenerationStrategy` originally resided in `benchmark`, because I was unsure of how useful it would be beyond benchmarking. Now that we are planning on using it for AutoML and the service API, we're moving it to `generator`. The kwargs filtering function is also not used (and probably shouldn't be) outside of `GenerationStrategy`, so moving it in there and making it private.

Reviewed By: amurthy1

Differential Revision: D14407297

fbshipit-source-id: c085f2688d3b0524544ad89e5bc25d7161bb6c7a