1/n Renaming base_encoder/decoder to encoder/decoder
Summary:
Just a simple renaming.

Context: I originally named these classes BaseDecoder/Encoder because they were meant to be subclassed, but as I'm writing documentation I'm realizing that I want to enable users to do as much as possible without writing their own code. So instead of subclassing, you should be able to achieve the same functionality by creating a custom instance of the class. See next diff for that ...

Reviewed By: amurthy1

Differential Revision: D14461228

fbshipit-source-id: b1bca47362e256a19db46b579be22eecc4a28f73