Add support for MathJax to Ax site
Summary:
Both the Markdown-based main doc pages + the notebook tutorials now support MathJax. MathJax will be available as a resource for all Docusaurus pages (it's injected into HTML head).

There are some weird interactions between MathJax and Plotly, so MathJax will not work on pages that have both until we move to online Plotly (via CDN) instead of injecting full library. We need to do that anyway, so that's coming up. Plotly also messes with the alignemnt of the equation blocks, so I had to override that via `!important`. Ugly, but it works.

Plotly, unlike MathJax, is not in html head, but on a per-page basis.

Reviewed By: Balandat

Differential Revision: D14490827

fbshipit-source-id: d5242aa9e0b0eee3e3a64fae88e0b2bd59645407