Fixes to allow base URLs that are not root
Summary:
Changes I made locally to be able to host Ax on https://home.fburl.com/~kkashin/Ax/.

Most of changes are to make links work with a baseUrl that is not `/` (default in siteConfig still).

Another small change is adding a flag to the make_docs.sh script that builds a static version of the site as opposed to starting up a server via `yarn start`.

Reviewed By: Balandat

Differential Revision: D14586757

fbshipit-source-id: a07414922e564a3fc2ddfa2e028b993f3cc13e16