Handle mixed none/non-none out of design parameters
Summary:
We require arms to have all parameters specified.

If an arm (status quo, almost surely) is out of design, it can have parameters with "None" as the value.
Previously, we were stripping the nulls here, during conversion of observation data. However, this still leaves any parameters that ARE specified as non-null in the arm, while they may still be out of design--causing errors in the search space transforms.

This diff wipes all parameters for any arms which are at least partially out-of-design.

This diff introduces one nasty edge case: if there are two out-of-design arms, this will currently error, when we de-duplicate observation features. Punting on that for now! T42118430

Reviewed By: bletham

Differential Revision: D14587163

fbshipit-source-id: 2f92f0915d798f4e9e34c8d241e5657b6b34bf7a