Add a slice plot to Ax site [example]
Summary:
Ben, here's an example of how to add the slice plot. Please feel free to commandeer if you want to make changes, add other plots, etc. I took this from n75801.

In order to do this, you want to grab the JS output from the plot; easy way to do this is to use `plot_config_to_html` (see https://fburl.com/jf3vwav1), or alternatively can grab it by inspecting the DOM.

You want to put the JS (everything within the requires) in a separate JS file, as I did for the slice.js. You want to call the JS script at the *end* of the markdown file (it has to be at the end, otherwise the markdown parser will stop rendering with the last script tag).

Note that the ID of the plot in the newPlot function in the JS needs to match the div id wherever you actually want the plot positioned.

I haven't tried this with two JS scripts at the bottom and multiple plots, but hopefully it works -- let me know if there are problems.

Reviewed By: bletham

Differential Revision: D14638165

fbshipit-source-id: 1e777c6eed9f77b5de7b89b495d8eeb3528aea1f