Modular botorch model
Summary:
As we had discussed in various places, the current `BotorchModel` implementation is somewhat monolithic and hard to extend.

This diff changes the `BotorchModel` implementation to be more modular. Basically, the functionality is the same, but the following sub-tasks are broken out into a functional sub-api (its docs should get some additional love):
- get and fit a model
- predict from the fitted model
- get an acquisition function from a fitted model and a bunch of other stuff
- optimize an acquisition function

The constructor of `BotorchModel` now takes callables for each of these  sub-tasks, which makes it easier to customize these 4 sub-components individually.

The defaults are just the elements of `BotorchModel` pulled out into functions.

See nb in the test plan for how to use this with a custom model.

Reviewed By: bletham

Differential Revision: D14636538

fbshipit-source-id: cea5f796927d30c6eb2ae4757712924eb0b6e449