Move Service API away from `SimpleExperiment`
Summary: The core feature of a `SimpleExperiment` is that it's both synchronous and sequential –– for the n+1-st trial to be evaluated, n-th trial must've been evaluated already. This will not be the case for service API users most of the time, so we should move the service API to regular `Experiment`, since we are creating everything under the hood anyway.

Reviewed By: amurthy1

Differential Revision: D14748199

fbshipit-source-id: 6d34e62ac148462e49807080a1542b905aff0a4f