Porting over relevant parts of Hyperband
Summary: Copied over parts of Hyperband code needed for Ax implementation of Hyperband. Will later separately dedup with the existing code in AutoML, as there will need to be some restructuring on the AutoML side. Till, please let me know if you think any more parts of the code / documentation can be removed from the simple version of Hyperband we are including with Ax.

Reviewed By: till-varoquaux

Differential Revision: D14814153

fbshipit-source-id: 2113f24e9edc1f2effa5bf16beef179e142f7009