Support fetching data for a custom set of metrics
Summary:
In old AE, you could rerun an analyses with new objectives, constraints, tracking metrics, etc. The metrics we used in the analyses were the ones defined on the analysis run, not on the experiment.

This currently doesn't work in Ax -- we always fetch data for the metrics as they are defined on the experiment, and we ignore the ones defined on the analysis run (i.e. inputs.metrics).

This should be a simple fix -- we just need our experiment.fetch_data method to take in an optional metrics argument. If provided, fetch data for these metrics; if not, fall back to experiment defaults.

Let me know if this seems okay or if we want to handle this a different way in Ax.

Reviewed By: amurthy1

Differential Revision: D14853864

fbshipit-source-id: f5afef0e92139a5130adecce9d321c9fbdd10fb5