Fix rejection sampling
Summary:
Rejection sampling is broken, due to a shape mismatch.

Points are coming in as (d,).

These means the left-hand side of our comparison below ends up being a (k,) np array, while the right-hand side is a (k, 1) np array. So, each element is compared, yielding a (k x k) set of constraint satisfactions, rather than (k x1).

This expands points to be (d x 1) before they are used in the constraint evaluation, and fixes this.

Reviewed By: lena-kashtelyan

Differential Revision: D14904331

fbshipit-source-id: 3c02467038bb7a81f0aaef7ec58c093977e74a25