1/n Cleaning up / speeding up storage: avoid unnecessary copying of analysis ents
Summary:
Part 1 in a series of diffs to clean up and speed up storage.

In this diff, I'm greatly simplifying the saving of experiments with analysis schedulers / runs. The reason this is a little tricky is that, when we call save_experiment in the FB code, it's mostly just a wrapper around save_experiment in the OSS code. And the OSS code does not know about analysis schedulers / runs, and so it ends up dropping / deleting all of them.

To handle this previously, I was doing something hacky in the FB-only code to just copy over the analysis schedulers from the old experiment to the updated experiment.

This was *really* slow and also kind of messy. A better solution (that I guess I didn't think of before) is to tell the OSS code just to ignore the analysis schedulers/ runs fields. That way they won't get dropped.

*Plus* if we do this, we can enable lazy-loading for analysis schedulers and runs, which also shaves off a bunch of time.

(There's another flow for updating or adding analysis schedulers / runs, so that's not a concern here.)

Reviewed By: sdsingh

Differential Revision: D14892291

fbshipit-source-id: 7ed9e1bbe87b296a91f50d536583e024bbb4ec6d