3/n Cleaning up / speeding up storage: remove unnecessary validation
Summary:
Part 3 in a series of diffs to clean up and speed up storage.

This is hopefully not too controversial, but I'm removing some validation here. Namely, I'm no longer validating (on the storage layer) that generator runs are immutable, nor am I validating that arm weights don't change.

The reason for the GR validation change is 1) it was super slow, 2) the user-facing class doesn't easily allow these kinds of changes anyway (you can see in the tests I have to use all _ accessors to even test this), so it felt redundant to have on the storage layer.

The reason for the arm validation change is: the status quo arm weight *can* actually change. This was sneaking by the tests due to a corner case, but later diffs address that, so it started erroring. We could add a special case for this, but again, the user facing class doesn't easily allow you to change the weight of an arm anyway, so I'm not sure we need to have it on the storage layer.

Reviewed By: amurthy1

Differential Revision: D14892368

fbshipit-source-id: 2b11b911e0700121cdbd492a4ef8a0266acc60ea