Easier saving of custom metrics/runners to SQA
Summary:
Piggybacking on the previous diff, adding helper functions register_runner and register_metric to be used when you want to create new runners/metrics in a notebook and have saving work for them.

Had to refactor things a bit for the nth time ... hopefully slowly converging on a system that I like.

Reviewed By: amurthy1

Differential Revision: D14954126

fbshipit-source-id: 956f6e17bc92e26e6d33782af0cb47eed0bca913