Include update time in stored fitting time
Summary:
ModelBridge stores the time it takes for calls to fit and gen. This diff makes two changes to how this is done to support the new update functionality:
* Time spent in update is included in fitting time
* The fit time stored on each GR is now just the time since the last GR, not the total fit time. (Since there may have been calls to update in between).

The second one is because in the benchmarks we compute total fit time by iterating through the GRs and summing their fit times, so having the fit time there be the cumulative fit time would be difficult to work with there.

Reviewed By: lena-kashtelyan

Differential Revision: D15015179

fbshipit-source-id: 9ce6e888c10586efbe84dd84d20123c17719e07b