GenerationStrategy move switching models to method
Summary: This moves switching models to be a method in GenerationStrategy. The primary reason is because I would like to be able to easily prototype things that inject custom logic into the model switch, and this will make that easy. As a side effect, I do find the current logic a bit hard to parse since there is a lot of negation happening and I think this cleans things up. It should do the exact same thing as before.

Reviewed By: lena-kashtelyan

Differential Revision: D15019757

fbshipit-source-id: 47d45d5ba324fefa85452ddfde280cc27b9cab48