Fix search space to choice bug
Summary:
The SS to choice transform creates one choice parameter out of all arms in the search space.

However, if there's only one arm, we then get an error about creating a choice parameter with only one value. We should create a fixed parameter instead.

(This should rarely happen but is an edge case that came up with Darwin)

Reviewed By: sdsingh

Differential Revision: D15104525

fbshipit-source-id: 4a4698e2889fc9148fa12df97dcc3ffeea69a2c8