Handle tutorial directories (#33)
Summary:
If a tutorial has associated files (e.g. Human in the loop), our current structure doesn't handle that.
This adds a dir flag in the tutorials json that can be used to set up the correct structure, and offer a button for downloading the data + source + notebook as a tar.gz
Pull Request resolved: https://github.com/facebook/Ax/pull/33

Reviewed By: amurthy1

Differential Revision: D15135566

Pulled By: sdsingh

fbshipit-source-id: f3614eab327f8cf7af587da1d688e07c12162b45