Update benchmarking tutorial
Summary:
Makes some updates to the benchmarking tutorial. Some minor text changes to clarify a few things. Adds a proper reference for the Gramacy problem.

I made a handful of changes so that this would run in a reasonable amount of time. I reduced the noise levels so that we could do less reps, and set the number of reps to 5. I limited it to a single Bayes Opt method (default botorch), which we just manually construct the factory function for as an illustration of how one could plug in custom models. I reduced the number of BayesOpt iterations to 10, by having an initialization of 10 points and 20 total iterations.

It now takes about 20 minutes to run.

Reviewed By: lena-kashtelyan

Differential Revision: D15152898

fbshipit-source-id: 9b7f68021c4c6b71f1538305ab7981587687d660