Add recommended parallelism setting
Summary:
For external schedulers, we should suggest a parallelism setting, so that they do not run into the 'need more data before can generate more' error.

For those with less context: this idea came out of a meeting with Teng, where we thought that it would be helpful for systems that integrate with Ax service API, to be able to learn what the recommended maximum parallelism at different points in the optimization and use it to obtain trials accordingly. This recommended parallelism, however, can still be bypassed by setting `enforce_sequential_optimization` to False in `AxClient` instantiation.

Reviewed By: ldworkin

Differential Revision: D15199728

fbshipit-source-id: ce34bfe5215630f6444688847edf19ec714adc59