Fix for floating point equality
Summary: An issue arises when trying to save an experiment with generator runs twice in a row. Has to do with the fact that the fit_time/gen_time of the Python object has more precision than what is stored in the database. Our update logic detects that the values are different, and tries to perform an update. But the update is a no-op because nothing changes in the db. SQLAlchemy doesn't like that.

Reviewed By: amurthy1

Differential Revision: D15342552

fbshipit-source-id: dda0e9ae86fc13f15448764164e25002121a5283