2/x AQE-specific report template
Summary:
Curious to get some thoughts on this diff. The context is: for AQE-import reports (which will be shown in the QE UI) we'd like to use QE-specific language, e.g. "group" instead of "arm", and maybe "goal" instead of "objective", "guardrail" instead of "constraint", etc.

One lazy approach: just do a find/replace on the report HTML. This feels dangerous (e.g. "harmful" would be come "hgroupful"). But maybe it's okay.

A better(?) approach: parameterize our text, as I do in this diff. This is pretty tedious, and there's no guarantee that I got everything. Plus, it introduces a really annoying "arm_noun" parameter in a million places.

Thoughts? See inline comments as well.

Reviewed By: amurthy1

Differential Revision: D15361058

fbshipit-source-id: 2c3df18f25d6177e3ee0302f9beba95450e964e3