Ad versioning logic to setup.py
Summary:
We want to add version information to Ax. We would like to use GitHub tags as the source of truth for the version (and only have a single source of truth) as opposed to check-in a hardcoded version file. This way, when releasing a new version, one can just tag a new commit on GitHub and not worry about bumping the version in other places.

See code comments and inline comments.

Reviewed By: kkashin

Differential Revision: D15556708

fbshipit-source-id: b86cb15197e15c5307ba8c0e9767a7597c1fa305