Test against botorch + gpytorch master (#105)
Summary:
We develop against gpytorch + botorch master, which means that tests will fail if we run them against the latest released version. This changes the CI settings in travis to use the most recent master for now, until we have a more stable base to work from.
Pull Request resolved: https://github.com/facebook/Ax/pull/105

Reviewed By: danielrjiang

Differential Revision: D15587607

Pulled By: Balandat

fbshipit-source-id: accf59e0c7389a5c0a7cd55477d066ed6c17737a