Add relative GP factory
Summary: There are some use cases where we want to fit a model multiple trials of data but also do relativization within each trial. Our current ModelBridge only supports doing relativization of all data against a single status quo, so we made a factory which relativizes, flips the outcome constraints to be non-relative, then passes it all to the GP. I tried doing this with transforms but it became too complicated.

Reviewed By: bletham

Differential Revision: D15620393

fbshipit-source-id: 7316d94f8af2e8efbf36ae95938c4f9714fcc383