Fix timeout on experiment loading
Summary:
So this issue we're trying to solve: we want to be able to load an object from the DB (within a session), but then access and manipulate it outside of the session (so that we don't get a timeout while performing operations that don't require a session).

SQLAlchemy doesn't make this easy. On commit, it automatically "expires" objects, which essentially makes them unusable outside of the session (https://stackoverflow.com/a/50272761).

To fix this, we turn off this functionality off by setting expire_on_commit = False.

Once we have this, we can separate out the logic that needs the session from the logic that just needs the session objects.

Reviewed By: kkashin

Differential Revision: D15743568

fbshipit-source-id: b005e256d60dfb4e0ffd9ee1af786131b52581fc