1/n Fetch precomp data for use in reports
Summary:
We currently have some functionality for fetching data from precomp, see https://fburl.com/rmgui2k9.

However, the dataframe returned by this function is not compatible with AE reports. See https://fb.quip.com/6RqcA4dnLZnt for a discussion of the differences.

As Matt suggested in that quip, to get the data we want, we should deltoid_stats table rather than deltoid_aggregations.

Here I'm adding a new method that fetches from that table and formats the data to be used in our reports.

Reviewed By: amurthy1

Differential Revision: D15781655

fbshipit-source-id: 0c84122f4d266254ff3c13ad6f9d1d834b10583b