Fixes to auto website deployment
Summary:
This should fix issues in https://travis-ci.com/facebook/Ax/jobs/207552057.

The sed issue stems from different semantics of sed across Linux and OSX. It's pretty hard to get this to work universally; right now, we're going to change this so that it works on Linux (in Travis). See https://stackoverflow.com/questions/4247068/sed-command-with-i-option-failing-on-mac-but-works-on-linux.

For manual running of the website versioning script on OSX (used for bumping up the version of the site and a new release), we should be able to install gnu-sed via `brew install gnu-sed --default-names`. See http://conqueringthecommandline.com/book/sed#cid57. We could also eventually just replace sed with Python or Perl for consistent behavior (recommendation here: https://stackoverflow.com/a/4247319/1191495).

Reviewed By: JoelMarcey

Differential Revision: D15899543

fbshipit-source-id: b14f653c75c4e664b2ad69b50131126e71a469f5