Save transforms
Summary: This is the first diff in the stack that makes the service API interruptible via making generation strategy and models serializable. We achieve the serialization of models through saving a model key and model kwargs on the generator run, and this will require being able to save transforms. These transforms are initialized in the ModelBridge, so we need to save transform classes and not instances.

Reviewed By: ldworkin

Differential Revision: D15564217

fbshipit-source-id: ec4d23509cf2fc65fb7c746a78e069ce68c9aa80