Extensible MNIST path for tutorials (#116)
Summary:
Richard from Ray submitted this PR; this helps customize the download path for MNIST and is especially useful for the Ax + Ray tutorial, because there MNIST needs to be downloaded or found on every trial (since Ray is for distrubuted training), so specifying a path would prevent it from being redownloaded on every trial in the tutorial setting.

This PR allows `load_mnist` to take in a data path to download the data.
Pull Request resolved: https://github.com/facebook/Ax/pull/116

Reviewed By: ldworkin

Differential Revision: D15901238

Pulled By: lena-kashtelyan

fbshipit-source-id: b5bde0bbeed8f1adf7aeaa333317c92a7ef5aa9f