Save model and kwargs on the generator run
Summary:
This is the second diff in the stack that makes the service API interruptible via making generation strategy and models serializable. We achieve this by saving model key and kwargs on the generator run, which also has a benefit of allowing us to know which model was used to produce a given generator run.

This functionality is used later in the stack, where I introduce the `Models` enum, which calls `ModelBridge. _set_kwargs_to_save`.

Reviewed By: sdsingh

Differential Revision: D15564947

fbshipit-source-id: 24ce6e3c5b28404476094360b1bef52cb2407601