Remove the deprecation warnings about factory functions for now, since they seem to clutter benchmarking
Summary: Expand the message in generation strategy instead, since that is where it really matters that the `Models` enum is used, since that will allow the generation strategy to be serialized. When we actually deprecate the factory functions (if we do), we can have them throw a `NonImplementedError` instead, directing users to the `Models` enum.

Reviewed By: Balandat

Differential Revision: D16223640

fbshipit-source-id: 9d00b7d6b95caabf05784b7746bd58e52be2f4e8