Add metric definitions to FO
Summary:
Add metric definitions for FO online, offline metric.

In the previous world, we had 1 metric class https://fburl.com/ynrv6gwt with a bunch of if branches based on whether the trial used an online or offline runner. Now we have 2 separate classes and the metric will track separate metric instances for the different trial types.

Previously, the name for the metric was one of the "canonical name"s in this config https://fburl.com/gzdt16hh which specifies how to fetch data for online, offline data for each canonical name

Our initial plan was that in the new world, we'd use the proper online/offline names (e.g. https://fburl.com/9waoz98s) to be the AE name for the respective metrics. However, I now realize this won't work because sometimes they're the same, which will cause name collisions in exp.metrics map.

I've opted for a bit hackier solution where we still use the canonical name for online and then just append '_ovm' to the offline metric definition. This way the final data still has a different name for online vs. offline measurements and it's still easy to lookup the definition in the config.

Reviewed By: ldworkin

Differential Revision: D16227997

fbshipit-source-id: a98b62ed2a1f89f45b2dff5e80ca799c4c5790f0