Propogate fixed features deep into plotting code to make it work for FO
Summary:
What a journey. To make predictions for FO, you need to specify whether to make the predictions in the "online" or "offline" domain. The problem is our model.predict happens very deep into our plot code so we need to pass this very deeply into the stack.

That's mostly what this diff does plus some other minor cleanup to get things working

Reviewed By: bletham

Differential Revision: D16456098

fbshipit-source-id: 100385f47b8cf5b7af0717a27bf3f5db0764f25f