Store experiment on generation strategy
Summary: To restore generation strategy from a serialized / stored state, an experiment is needed (for reinstantiating the model that was instantiated on the strategy). For to be able to store a reference to the experiment, we add a field on generation strategy, to record which experiment it has been used with.

Reviewed By: ldworkin

Differential Revision: D16604356

fbshipit-source-id: 47649e2f3dda4c2e2042a60b93855294b301294b