Execute Ax tutorials when building site
Summary:
Previous behavior when building Ax site was to convert Ax tutorials from ipynb to html, without executing the tutorial (thus relying on manual execution of checked in tutorial).

This led to several problems:
a) Tutorials not executed completely / missing outputs
b) No thorough testing of tutorials, meaning that sometimes they may not run e2e

This diff adds support for executing tutorials as part of website building (and also tracking execution time).

For the moment, I'm try / catching tutorial execution, and falling back on conversion from checked-in ipynb to html directly. This is temporary and the goal is to remove that once the tutorials that currently don't execute are fixed.

Right now, plan is to run this on each commit to master on Travis & when releasing a new version of Ax & building the site locally. If that get cumbersome to build every time, we can build the site nightly on Travis via a cron job (and conditional builds, see https://docs.travis-ci.com/user/conditional-builds-stages-jobs/, https://docs.travis-ci.com/user/conditions-v1).

Reviewed By: ldworkin

Differential Revision: D16641459

fbshipit-source-id: 12ba1e0952bb120500af3d95a98f4c232b3ada67