Check for an MTGP model with TrialAsTask transform
Summary:
Check if the model have 'TrialAsTask' transform to verify that it is a multi-task model.
In this case we belive it is a MTGP and the model can have  observations of arms with different features but same name.

And we don't need to show the error for this multi-task models.

Reviewed By: sdsingh

Differential Revision: D16660238

fbshipit-source-id: 43228a3258820664029c921b403056600a88ce45