Getting rid of "cannot change time_created" error
Summary:
Guys, it was real. This error has haunted us for a long time, but NO MORE.

(In my defense, I still think it was kinda the right behavior, but it is clearly *way* more trouble than it's worth!)

If you re-create an experiment with the same name, we'll just use the original time_created value.

Reviewed By: stevemandala

Differential Revision: D16717161

fbshipit-source-id: b56b67d12d294b5473b42c05459ddae1a8cc241e