Make joint_/sequential_optimize return acq function values
Summary:
This is useful both for diagnostics as well as making initialization strategies less cumbersome to write (in particular if `return_best_only=False`, and one would like to re-compute the acquisition function values - right now, this means re-evaluating the acquisition function unnecessarily, doing double work).

This means a breaking change, so we'll have to be careful to announce this in a transparent fashion and synchronize the release with Ax. I guess we can lump that into the breaking change planned for the output dimensions.

I may also consider soft-deprecating `joint_optimize` and `sequential_optimize` and instead have a single `optimize_acqf` with a `sequential` flag.

Finally, we may consider getting rid / deprecating of `get_best_candidates` since this is a one-liner....

Differential Revision: D16695425

fbshipit-source-id: ab08d458a930e064992d80fa83da2ce4680a7822