Make joint_/sequential_optimize return acq function values (#149)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/149

Pull Request resolved: https://github.com/pytorch/botorch/pull/234

This is useful both for diagnostics as well as making initialization strategies less cumbersome to write (in particular if `return_best_only=False`, and one would like to re-compute the acquisition function values - right now, this means re-evaluating the acquisition function unnecessarily, doing double work).

This means a breaking change, so we'll have to be careful to announce this in a transparent fashion and synchronize the release with Ax.

We may consider getting rid / deprecating of `get_best_candidates` since this is a one-liner....

Reviewed By: sdaulton

Differential Revision: D16695425

fbshipit-source-id: 763a82e93b30e9df24bc7f69d9d31e5ae6016033