Cap parameter transform
Summary:
For the initialization of the downsampling searcher in AutoML, we need to cap the downsampling parameter for the Sobol phase, since there is no reason to believe that a high fidelity on the quasi-random points is necessary. As we add the KG searcher into open-source, this transform will likely come in handy outside of AutoML, too. What do people think?

I started with a very basic version, which we can extend if need be.

Reviewed By: bkarrer

Differential Revision: D16698092

fbshipit-source-id: b16e7808450e541bf18a3668cf19c4e6f52a43b9