Chose status quo using trial index
Summary:
Add trial index to determine which of trials' status quo to use for the experiment. In the get_MTGP function we select the corresponding status quo and pass it in the status_quo_features to the ModelBridge.

If it is None and there are more than one value of status quo for the experiment, the last of the existed values will be chosen.

It is used for plotting.

Reviewed By: sdsingh

Differential Revision: D16673399

fbshipit-source-id: 3636652a5f2c1ed2b8c51e4b05b960cdd68a5641