Store state of stateful models, to add to kwargs when reinstantiating
Summary: Some of the Ax models (currently only `SobolGenerator`, but in the future Hyperband & others) have state, so it's not enough for those to record the kwargs, with which they've been instantiated. This diff introduces the logic used to store the state of those models, such that they can be reinstantiated from a generator run they produced.

Reviewed By: stevemandala

Differential Revision: D16993233

fbshipit-source-id: 5e96ec993535399e63446016c11cb2892b204b62