Warm start model fitting (#156)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/156

Adds a warm_start_model_fit attribute to `BotorchModel` to extract the state_dict from a previous model & pass it to get_and_fit_model along with a keyword argument that indicates that the model should be refit.

get_and_fit_model loads the state_dict if provided and refits the model when specified.

Reviewed By: sdaulton

Differential Revision: D16792879

fbshipit-source-id: 85137a287648142bd7cbde4c7af6924b3b03af52