Enable `get_pending_observation_features` for `BatchTrial`-s
Summary:
Expanding `get_pending_observation_features` to also get the pending features from batched trials. This is used in Maximus, which uses ax batch client, subsequently in the stack.

Note: right now this will only be used in Maximus; the best reviewer for this would be Ben, but since he is away and Maximus depends on this change, I think we can get this in, and I will get Ben's approval when he is back.

Reviewed By: ldworkin, stevemandala

Differential Revision: D17102905

fbshipit-source-id: ced2524265ab05e6fe9e0a755670390dc500ad97